requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
SQLAlchemy==2.0.41
psycopg2-binary==2.9.10
python-dotenv==1.0.0
ldap3==2.9.1
requests==2.31.0
msal==1.24.1
psutil
cryptography==42.0.5
tabulate==0.9.0
pyodbc==5.2.0

ruff
mypy
types-tabulate